requirements.txt
# Pharma Regulatory Pulse — Python Dependencies
# Pinned for reproducibility

# AI synthesis
openai>=1.50.0

# Primary news search (LLM-optimised, 93.3% accuracy on SimpleQA)
tavily-python>=0.5.0

# Email formatting
mjml>=0.8.0

# Email delivery
resend>=2.0.0

# Web / RSS research
feedparser>=6.0.0
requests>=2.32.0

# NLP deduplication
scikit-learn>=1.5.0
numpy>=1.26.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0
